Problem: Files not always fully read
Solution: Make size more dynamic.

We can either read from a command (./size in the directory) or we can go
the lazy and probably slower route: cat the command and just read from
the pipe until we can determine the size. We do this reading in 4 meg
chunks to try and read quickly without overwhelming a system with many
files that might be read at once.

Success size is >0. On failure, the file size is set to 0.

Also changed some Entries to be pointers to be properly referenced
instead of copying the values around.