Problem: Tar needs sizes resolved before starting
Solution: Add tracking of when sizes are computed.

As an output mechanism, I considered:
- Forking to background
	Dismissed because golang doesn't seem to really do that as
	a pattern.
- Setting an ACL or ATTR on size done
	Gotta poll, and would involve fitting a weird requirement in,
	since no ACL seems to say "Size information is ready" and
	somehow o-r doesn't shout "I'm ready to read." A working
	fallback solution if nothing else works.
- Do a named socket or something of that sort.
	And we got a winner. A named pipe. This can do blocking when cat
	reads the read end of a FIFO, so I don't need to poll in shell.
	It also supports a simple writer, so I can just trigger shell
	continuing with a single Write().

There may be a state bug here where it is possible to trigger a re-read
before your sizes are computed, and your sizes may abort instead of
ending. Or maybe it'll always complete. Wasn't worth figuring it out at
this point since re-reading the tree is not currently a feature.

No point making it perfectly robust in advance, right? This just happens
to fit the need and probably work into the future.